Add files via upload
Complete change of how the statistic works. Main program is now for technical replicates only. For the statistical analysis of biological replicates use the menu to navigate to the statistical analysis. The significance is now tested differently for paired/unpaired std. deviated/non std. deviated.

Plots can now be adjusted to be black and white (necessary for some publications)